Cleanup context creation
