+++ lib/server/createGraphQLMethod.js
@@ -1,14 +1,17 @@
 import { execute } from 'graphql';
-import { DEFAULT_CREATE_CONTEXT } from '../common/defaults';
+import { contextToFunction } from './contextToFunction';
 
 export const DDP_APOLLO_SCHEMA_REQUIRED = 'DDP_APOLLO_SCHEMA_REQUIRED';
 
-export function createGraphQLMethod(schema, {
-  createContext = DEFAULT_CREATE_CONTEXT,
+export function createGraphQLMethod({
+  schema,
+  context,
 }) {
   if (!schema) {
     throw new Error(DDP_APOLLO_SCHEMA_REQUIRED);
   }
+
+  const createContext = contextToFunction(context);
 
   return async function graphQlMethod({ query, variables, operationName } = {}, clientContext) {
     if (!query) {
@@ -20,13 +23,13 @@ export function createGraphQLMethod(schema, {
     }
 
     const { userId } = this;
-    const context = await createContext({ userId }, clientContext);
+    const completeContext = await createContext({ userId }, clientContext);
 
     return execute(
       schema,
       query,
       {},
-      context,
+      completeContext,
       variables,
       operationName,
     );
