+++ lib/server/createGraphQLPublication.js
@@ -1,9 +1,9 @@
 import { Meteor } from 'meteor/meteor';
 import { subscribe } from 'graphql';
 import forAwaitEach from '../common/forAwaitEach';
+import { contextToFunction } from './contextToFunction';
 import {
   DEFAULT_PUBLICATION,
-  DEFAULT_CREATE_CONTEXT,
   GRAPHQL_SUBSCRIPTION_MESSAGE_TYPE,
 } from '../common/defaults';
 
@@ -12,13 +12,15 @@ const SUBSCRIPTION_MESSAGE_TYPE = 'pong';
 
 export function createGraphQLPublication({
   schema,
+  context,
   publication = DEFAULT_PUBLICATION,
-  createContext = DEFAULT_CREATE_CONTEXT,
 } = {}) {
   if (!subscribe) {
     console.warn('DDP-Apollo: You need graphl@0.11 or higher for subscription support');
     return;
   }
+
+  const createContext = contextToFunction(context);
 
   Meteor.publish(publication, function publishGraphQL({
     query,
@@ -37,13 +39,13 @@ export function createGraphQLPublication({
     }
 
     const subscribeToQuery = async () => {
-      const context = await createContext({ userId }, clientContext);
+      const completeContext = await createContext({ userId }, clientContext);
 
       const iterator = await subscribe(
         schema,
         query,
         {},
-        context,
+        completeContext,
         variables,
         operationName,
       );
