+++ lib/server/setup.js
@@ -1,23 +1,9 @@
 import { Meteor } from 'meteor/meteor';
 import { createGraphQLMethod } from './createGraphQLMethod';
 import { createGraphQLPublication } from './createGraphQLPublication';
-import {
-  DEFAULT_METHOD,
-  DEFAULT_CREATE_CONTEXT,
-} from '../common/defaults';
+import { DEFAULT_METHOD } from '../common/defaults';
 
 export const DDP_APOLLO_SCHEMA_REQUIRED = 'DDP_APOLLO_SCHEMA_REQUIRED';
-
-function contextToFunction(context) {
-  switch (typeof context) {
-    case 'object':
-      return defaultContext => ({ ...defaultContext, ...context });
-    case 'function':
-      return context;
-    default:
-      return DEFAULT_CREATE_CONTEXT;
-  }
-}
 
 export function setup({
   schema,
@@ -29,15 +15,16 @@ export function setup({
     throw new Error(DDP_APOLLO_SCHEMA_REQUIRED);
   }
 
-  const createContext = contextToFunction(context);
-
   Meteor.methods({
-    [method]: createGraphQLMethod(schema, { createContext }),
+    [method]: createGraphQLMethod({
+      schema,
+      context,
+    }),
   });
 
   createGraphQLPublication({
     schema,
-    createContext,
     publication,
+    context,
   });
 }
